split to monitor module
