+++ docs/coderef/flowdyn.monitors.md
@@ -1,0 +1,6 @@
+:::flowdyn.monitors
+    selection:
+        filters:
+            - "!.*TiTs.*"
+            - "!.*PiPs.*"
+            - "!.*MachTi"
